+++ nipype/pipeline/engine/tests/test_utils.py
@@ -10,14 +10,35 @@ from builtins import range, open
 import os
 import sys
 from copy import deepcopy
-from shutil import rmtree
 import pytest
 
 from ... import engine as pe
 from ....interfaces import base as nib
 from ....interfaces import utility as niu
 from .... import config
-from ..utils import merge_dict, clean_working_directory, write_workflow_prov
+from ..utils import clean_working_directory, write_workflow_prov
+
+
+class InputSpec(nib.TraitedSpec):
+    in_file = nib.File(exists=True, copyfile=True)
+
+
+class OutputSpec(nib.TraitedSpec):
+    output1 = nib.traits.List(nib.traits.Int, desc='outputs')
+
+
+class UtilsTestInterface(nib.BaseInterface):
+    input_spec = InputSpec
+    output_spec = OutputSpec
+
+    def _run_interface(self, runtime):
+        runtime.returncode = 0
+        return runtime
+
+    def _list_outputs(self):
+        outputs = self._outputs().get()
+        outputs['output1'] = [1]
+        return outputs
 
 
 def test_identitynode_removal(tmpdir):
@@ -99,204 +120,11 @@ def test_clean_working_directory(tmpdir):
     config.set_default_config()
 
 
-def test_outputs_removal(tmpdir):
-    def test_function(arg1):
-        import os
-        file1 = os.path.join(os.getcwd(), 'file1.txt')
-        file2 = os.path.join(os.getcwd(), 'file2.txt')
-        fp = open(file1, 'wt')
-        fp.write('%d' % arg1)
-        fp.close()
-        fp = open(file2, 'wt')
-        fp.write('%d' % arg1)
-        fp.close()
-        return file1, file2
-
-    n1 = pe.Node(
-        niu.Function(
-            input_names=['arg1'],
-            output_names=['file1', 'file2'],
-            function=test_function),
-        base_dir=tmpdir.strpath,
-        name='testoutputs')
-    n1.inputs.arg1 = 1
-    n1.config = {'execution': {'remove_unnecessary_outputs': True}}
-    n1.config = merge_dict(deepcopy(config._sections), n1.config)
-    n1.run()
-    assert tmpdir.join(n1.name, 'file1.txt').check()
-    assert tmpdir.join(n1.name, 'file1.txt').check()
-    n1.needed_outputs = ['file2']
-    n1.run()
-    assert not tmpdir.join(n1.name, 'file1.txt').check()
-    assert tmpdir.join(n1.name, 'file2.txt').check()
-
-
-class InputSpec(nib.TraitedSpec):
-    in_file = nib.File(exists=True, copyfile=True)
-
-
-class OutputSpec(nib.TraitedSpec):
-    output1 = nib.traits.List(nib.traits.Int, desc='outputs')
-
-
-class UtilsTestInterface(nib.BaseInterface):
-    input_spec = InputSpec
-    output_spec = OutputSpec
-
-    def _run_interface(self, runtime):
-        runtime.returncode = 0
-        return runtime
-
-    def _list_outputs(self):
-        outputs = self._outputs().get()
-        outputs['output1'] = [1]
-        return outputs
-
-
-def test_inputs_removal(tmpdir):
-    file1 = tmpdir.join('file1.txt')
-    file1.write('dummy_file')
-    n1 = pe.Node(
-        UtilsTestInterface(), base_dir=tmpdir.strpath, name='testinputs')
-    n1.inputs.in_file = file1.strpath
-    n1.config = {'execution': {'keep_inputs': True}}
-    n1.config = merge_dict(deepcopy(config._sections), n1.config)
-    n1.run()
-    assert tmpdir.join(n1.name, 'file1.txt').check()
-    n1.inputs.in_file = file1.strpath
-    n1.config = {'execution': {'keep_inputs': False}}
-    n1.config = merge_dict(deepcopy(config._sections), n1.config)
-    n1.overwrite = True
-    n1.run()
-    assert not tmpdir.join(n1.name, 'file1.txt').check()
-
-
-def test_outputs_removal_wf(tmpdir):
-    def test_function(arg1):
-        import os
-        file1 = os.path.join(os.getcwd(), 'file1.txt')
-        file2 = os.path.join(os.getcwd(), 'file2.txt')
-        file3 = os.path.join(os.getcwd(), 'file3.txt')
-        file4 = os.path.join(os.getcwd(), 'subdir', 'file1.txt')
-        files = [file1, file2, file3, file4]
-        os.mkdir("subdir")
-        for filename in files:
-            with open(filename, 'wt') as fp:
-                fp.write('%d' % arg1)
-        return file1, file2, os.path.join(os.getcwd(), "subdir")
-
-    def test_function2(in_file, arg):
-        import os
-        in_arg = open(in_file).read()
-        file1 = os.path.join(os.getcwd(), 'file1.txt')
-        file2 = os.path.join(os.getcwd(), 'file2.txt')
-        file3 = os.path.join(os.getcwd(), 'file3.txt')
-        files = [file1, file2, file3]
-        for filename in files:
-            with open(filename, 'wt') as fp:
-                fp.write('%d' % arg + in_arg)
-        return file1, file2, 1
-
-    def test_function3(arg):
-        import os
-        return arg
-
-    for plugin in ('Linear', ):  # , 'MultiProc'):
-        n1 = pe.Node(
-            niu.Function(
-                input_names=['arg1'],
-                output_names=['out_file1', 'out_file2', 'dir'],
-                function=test_function),
-            name='n1',
-            base_dir=tmpdir.strpath)
-        n1.inputs.arg1 = 1
-
-        n2 = pe.Node(
-            niu.Function(
-                input_names=['in_file', 'arg'],
-                output_names=['out_file1', 'out_file2', 'n'],
-                function=test_function2),
-            name='n2',
-            base_dir=tmpdir.strpath)
-        n2.inputs.arg = 2
-
-        n3 = pe.Node(
-            niu.Function(
-                input_names=['arg'],
-                output_names=['n'],
-                function=test_function3),
-            name='n3',
-            base_dir=tmpdir.strpath)
-
-        wf = pe.Workflow(
-            name="node_rem_test" + plugin, base_dir=tmpdir.strpath)
-        wf.connect(n1, "out_file1", n2, "in_file")
-
-        wf.run(plugin='Linear')
-
-        for remove_unnecessary_outputs in [True, False]:
-            config.set_default_config()
-            wf.config = {
-                'execution': {
-                    'remove_unnecessary_outputs': remove_unnecessary_outputs
-                }
-            }
-            rmtree(os.path.join(wf.base_dir, wf.name))
-            wf.run(plugin=plugin)
-
-            assert os.path.exists(
-                os.path.join(wf.base_dir, wf.name, n1.name,
-                             'file2.txt')) != remove_unnecessary_outputs
-            assert os.path.exists(
-                os.path.join(wf.base_dir, wf.name, n1.name, "subdir",
-                             'file1.txt')) != remove_unnecessary_outputs
-            assert os.path.exists(
-                os.path.join(wf.base_dir, wf.name, n1.name, 'file1.txt'))
-            assert os.path.exists(
-                os.path.join(wf.base_dir, wf.name, n1.name,
-                             'file3.txt')) != remove_unnecessary_outputs
-            assert os.path.exists(
-                os.path.join(wf.base_dir, wf.name, n2.name, 'file1.txt'))
-            assert os.path.exists(
-                os.path.join(wf.base_dir, wf.name, n2.name, 'file2.txt'))
-            assert os.path.exists(
-                os.path.join(wf.base_dir, wf.name, n2.name,
-                             'file3.txt')) != remove_unnecessary_outputs
-
-        n4 = pe.Node(UtilsTestInterface(), name='n4', base_dir=tmpdir.strpath)
-        wf.connect(n2, "out_file1", n4, "in_file")
-
-        def pick_first(l):
-            return l[0]
-
-        wf.connect(n4, ("output1", pick_first), n3, "arg")
-        for remove_unnecessary_outputs in [True, False]:
-            for keep_inputs in [True, False]:
-                config.set_default_config()
-                wf.config = {
-                    'execution': {
-                        'keep_inputs': keep_inputs,
-                        'remove_unnecessary_outputs':
-                        remove_unnecessary_outputs
-                    }
-                }
-                rmtree(os.path.join(wf.base_dir, wf.name))
-                wf.run(plugin=plugin)
-                assert os.path.exists(
-                    os.path.join(wf.base_dir, wf.name, n2.name, 'file1.txt'))
-                assert os.path.exists(
-                    os.path.join(wf.base_dir, wf.name, n2.name,
-                                 'file2.txt')) != remove_unnecessary_outputs
-                assert os.path.exists(
-                    os.path.join(wf.base_dir, wf.name, n4.name,
-                                 'file1.txt')) == keep_inputs
-
-
-def fwhm(fwhm):
-    return fwhm
-
-
 def create_wf(name):
+    """Creates a workflow for the following tests"""
+    def fwhm(fwhm):
+        return fwhm
+
     pipe = pe.Workflow(name=name)
     process = pe.Node(
         niu.Function(
