+++ nipype/pipeline/engine/tests/test_workflows.py
@@ -3,12 +3,17 @@
 # vi: set ft=python sts=4 ts=4 sw=4 et:
 """Tests for the engine workflows module
 """
+import os
+from shutil import rmtree
+from itertools import product
 import pytest
 import networkx as nx
 
-from ... import engine as pe
+from .... import config
 from ....interfaces import utility as niu
+from ... import engine as pe
 from .test_base import EngineTestInterface
+from .test_utils import UtilsTestInterface
 
 
 def test_init():
@@ -68,16 +73,14 @@ def test_doubleconnect():
     b = pe.Node(niu.IdentityInterface(fields=['a', 'b']), name='b')
     flow1 = pe.Workflow(name='test')
     flow1.connect(a, 'a', b, 'a')
-    x = lambda: flow1.connect(a, 'b', b, 'a')
     with pytest.raises(Exception) as excinfo:
-        x()
+        flow1.connect(a, 'b', b, 'a')
     assert "Trying to connect" in str(excinfo.value)
 
     c = pe.Node(niu.IdentityInterface(fields=['a', 'b']), name='c')
     flow1 = pe.Workflow(name='test2')
-    x = lambda: flow1.connect([(a, c, [('b', 'b')]), (b, c, [('a', 'b')])])
     with pytest.raises(Exception) as excinfo:
-        x()
+        flow1.connect([(a, c, [('b', 'b')]), (b, c, [('a', 'b')])])
     assert "Trying to connect" in str(excinfo.value)
 
 
@@ -106,3 +109,123 @@ def test_duplicate_node_check():
     with pytest.raises(IOError) as excinfo:
         wf.connect(wf_connections)
     assert 'Duplicate node name "selector3" found.' == str(excinfo.value)
+
+
+def _test_function(arg1):
+    import os
+    file1 = os.path.join(os.getcwd(), 'file1.txt')
+    file2 = os.path.join(os.getcwd(), 'file2.txt')
+    file3 = os.path.join(os.getcwd(), 'file3.txt')
+    file4 = os.path.join(os.getcwd(), 'subdir', 'file4.txt')
+    os.mkdir("subdir")
+    for filename in [file1, file2, file3, file4]:
+        with open(filename, 'wt') as fp:
+            fp.write('%d' % arg1)
+    return file1, file2, os.path.join(os.getcwd(), "subdir")
+
+
+def _test_function2(in_file, arg):
+    import os
+    with open(in_file, 'rt') as fp:
+        in_arg = fp.read()
+
+    file1 = os.path.join(os.getcwd(), 'file1.txt')
+    file2 = os.path.join(os.getcwd(), 'file2.txt')
+    file3 = os.path.join(os.getcwd(), 'file3.txt')
+    files = [file1, file2, file3]
+    for filename in files:
+        with open(filename, 'wt') as fp:
+            fp.write('%d' % arg + in_arg)
+    return file1, file2, 1
+
+
+def _test_function3(arg):
+    return arg
+
+
+@pytest.mark.parametrize(
+    'plugin, remove_unnecessary_outputs, keep_inputs',
+    list(product(['Linear', 'MultiProc'], [False, True], [True, False])))
+def test_outputs_removal_wf(tmpdir, plugin, remove_unnecessary_outputs,
+                            keep_inputs):
+    config.set_default_config()
+    config.set('execution', 'remove_unnecessary_outputs',
+               remove_unnecessary_outputs)
+    config.set('execution', 'keep_inputs', keep_inputs)
+
+    n1 = pe.Node(
+        niu.Function(
+            output_names=['out_file1', 'out_file2', 'dir'],
+            function=_test_function),
+        name='n1',
+        base_dir=tmpdir.strpath)
+    n1.inputs.arg1 = 1
+
+    n2 = pe.Node(
+        niu.Function(
+            output_names=['out_file1', 'out_file2', 'n'],
+            function=_test_function2),
+        name='n2',
+        base_dir=tmpdir.strpath)
+    n2.inputs.arg = 2
+
+    n3 = pe.Node(
+        niu.Function(
+            output_names=['n'],
+            function=_test_function3),
+        name='n3',
+        base_dir=tmpdir.strpath)
+
+    wf = pe.Workflow(
+        name="node_rem_test" + plugin, base_dir=tmpdir.strpath)
+
+    wf.connect(n1, "out_file1", n2, "in_file")
+    wf.run(plugin=plugin)
+
+    # Necessary outputs HAVE to exist
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n1.name, 'file1.txt'))
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n2.name, 'file1.txt'))
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n2.name, 'file2.txt'))
+
+    # Unnecessary outputs exist only iff remove_unnecessary_outputs is True
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n1.name,
+                     'file2.txt')) is not remove_unnecessary_outputs
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n1.name, "subdir",
+                     'file4.txt')) is not remove_unnecessary_outputs
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n1.name,
+                     'file3.txt')) is not remove_unnecessary_outputs
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n2.name,
+                     'file3.txt')) is not remove_unnecessary_outputs
+
+    n4 = pe.Node(UtilsTestInterface(), name='n4', base_dir=tmpdir.strpath)
+    wf.connect(n2, "out_file1", n4, "in_file")
+
+    def pick_first(l):
+        return l[0]
+
+    wf.connect(n4, ("output1", pick_first), n3, "arg")
+    rmtree(os.path.join(wf.base_dir, wf.name))
+    wf.run(plugin=plugin)
+
+    # Test necessary outputs
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n2.name, 'file1.txt'))
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n2.name, 'file1.txt'))
+
+    # Test unnecessary outputs
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n2.name,
+                     'file2.txt')) is not remove_unnecessary_outputs
+
+    # Test keep_inputs
+    assert os.path.exists(
+        os.path.join(wf.base_dir, wf.name, n4.name,
+                     'file1.txt')) is keep_inputs
